+++ src/app/shared/module/poe/component/item-frame-value/item-frame-value.component.ts
@@ -35,29 +35,35 @@ export class ItemFrameValueComponent implements OnInit {
 
   public onMouseUp(event: MouseEvent): void {
     event.stopImmediatePropagation();
+    /* tslint:disable */
     if (event.which === 2) {
       this.resetValue(true, true);
     } else if (event.which === 3) {
       this.toggleValue(true, true);
     }
+    /* tslint:enable */
   }
 
   public onMouseUpMin(event: MouseEvent): void {
     event.stopImmediatePropagation();
+    /* tslint:disable */
     if (event.which === 2) {
       this.resetValue(true, false);
     } else if (event.which === 3) {
       this.toggleValue(true, false);
     }
+    /* tslint:enable */
   }
 
   public onMouseUpMax(event: MouseEvent): void {
     event.stopImmediatePropagation();
+    /* tslint:disable */
     if (event.which === 2) {
       this.resetValue(false, true);
     } else if (event.which === 3) {
       this.toggleValue(false, true);
     }
+    /* tslint:enable */
   }
 
   public onWheel(event: WheelEvent): void {