+++ src/app/shared/module/poe/service/item/query/item-search-filters-stats.service.ts
@@ -25,23 +25,45 @@ export class ItemSearchFiltersStatsService implements ItemSearchFiltersService {
                     const mins = stat.values.filter(x => x.min !== undefined);
                     const maxs = stat.values.filter(x => x.max !== undefined);
 
-                    let min = mins.reduce((a, b) => a + b.min, 0);
-                    let max = maxs.reduce((a, b) => a + b.max, 0);
-
-                    if (mins.length > 0 && maxs.length > 0) {
-                        min = Math.floor(min / mins.length);
-                        max = Math.ceil(max / maxs.length);
-                        filter.value = {
-                            min: min > max ? max : min,
-                            max: max > min ? max : min
-                        };
-                    } else if (mins.length > 0) {
-                        min = Math.floor(min / mins.length);
-                        filter.value = { min };
-                    } else if (maxs.length > 0) {
-                        max = Math.ceil(max / maxs.length);
-                        filter.value = { max };
+                    if (mins.length === 0 && maxs.length === 0) {
+                        return filter;
                     }
+
+                    const negate = !stat.negated && stat.predicate[0] === 'N'
+                        ? -1 : 1;
+
+                    let min;
+                    if (mins.length > 0) {
+                        min = Math.floor(mins.reduce((a, b) => a + b.min, 0) / mins.length) * negate;
+                    }
+
+                    let max;
+                    if (maxs.length > 0) {
+                        max = Math.floor(maxs.reduce((a, b) => a + b.max, 0) / maxs.length) * negate;
+                    }
+
+                    if (min !== undefined && max !== undefined) {
+                        if (min > max) {
+                            const tmp = min;
+                            min = max;
+                            max = tmp;
+                        }
+                    } else if (min !== undefined) {
+                        if (min >= 0) {
+                            max = 999;
+                        } else {
+                            max = -1;
+                        }
+                    } else if (max !== undefined) {
+                        if (max >= 0) {
+                            min = 0;
+                        } else {
+                            min = max;
+                            max = -1;
+                        }
+                    }
+
+                    filter.value = { min, max };
                     return filter;
                 })
             }